+++ packages/api2/src/database/interfaces/migration-context.interface.ts
@@ -1,6 +1,6 @@
 import { PeriodsService } from '@/periods/services/periods.service';
 import { PeriodSettingsService } from '@/periodsettings/periodsettings.service';
-import { PraiseService } from '@/praise/praise.service';
+import { PraiseService } from '@/praise/services/praise.service';
 import { QuantificationsService } from '@/quantifications/services/quantifications.service';
 import { SettingsService } from '@/settings/settings.service';
 import { UsersService } from '@/users/users.service';
