+++ packages/api2/src/praise/praise.controller.ts
@@ -25,7 +25,7 @@ import { isArray } from 'class-validator';
 import { Types } from 'mongoose';
 import { ObjectIdPipe } from '@/shared/pipes/object-id.pipe';
 import { QuantifyMultipleInputDto } from './dto/quantify-multiple-input.dto';
-import { PraiseService } from './praise.service';
+import { PraiseService } from './services/praise.service';
 import { Praise } from './schemas/praise.schema';
 import { PraisePaginatedQueryDto } from './dto/praise-paginated-query.dto';
 import { PermissionsGuard } from '@/auth/guards/permissions.guard';
@@ -38,6 +38,8 @@ import { PraisePaginatedResponseDto } from './dto/praise-paginated-response.dto'
 import { Response } from 'express';
 import { ExportInputDto } from '@/shared/dto/export-input.dto';
 import { allExportsDirPath } from '@/shared/fs.shared';
+import { optionsHash } from '@/shared/export.shared';
+import { PraiseExportService } from './services/praise-export.service';
 
 @Controller('praise')
 @ApiTags('Praise')
@@ -47,7 +49,10 @@ import { allExportsDirPath } from '@/shared/fs.shared';
 // @UseGuards(PermissionsGuard)
 // @UseGuards(JwtAuthGuard)
 export class PraiseController {
-  constructor(private readonly praiseService: PraiseService) {}
+  constructor(
+    private readonly praiseService: PraiseService,
+    private readonly praiseExportService: PraiseExportService,
+  ) {}
 
   @Get()
   @ApiOperation({ summary: 'List praise items, paginated results' })
@@ -80,23 +85,34 @@ export class PraiseController {
     @Res({ passthrough: true }) res: Response,
   ): Promise<StreamableFile> {
     const { format } = options;
-    const exportsRootDirName = `${allExportsDirPath}/praise`;
-    const exportDirName = await this.praiseService.getExportDirName();
-    const exportId = this.praiseService.getExportId(options);
-    const exportFilePath = `${exportsRootDirName}/${exportDirName}/${exportId}/praise.${format}`;
+    // Root path for all exports
+    const rootPath = `${allExportsDirPath}/praise`;
 
-    // Cached export don't exist, clear cache and generate new export
-    if (!fs.existsSync(exportFilePath)) {
-      if (!fs.existsSync(`${exportsRootDirName}/${exportDirName}`)) {
-        fs.rmSync(exportsRootDirName, { recursive: true, force: true });
+    // Directory level 1 is the latest praise id
+    const dirLevel1 = (await this.praiseService.findLatest())._id.toString();
+
+    // Directory level 2 is the hashed options
+    const dirLevel2 = optionsHash(options);
+
+    const dirPath = `${rootPath}/${dirLevel1}/${dirLevel2}`;
+    const filePath = `${dirPath}/praise.${format}`;
+
+    if (!fs.existsSync(filePath)) {
+      // If cached export don't exist
+      if (!fs.existsSync(`${rootPath}/${dirLevel1}`)) {
+        // If the latest praise id folder doesn't exist,
+        // database hase been updated, clear all cached exports
+        fs.rmSync(rootPath, { recursive: true, force: true });
       }
 
-      console.log("Export file doesn't exist, generating new export files");
+      // Create directory for new export
+      fs.mkdirSync(dirPath, { recursive: true });
+
       // Generate new export files
-      await this.praiseService.generateAllExports(options);
+      await this.praiseExportService.generateAllExports(dirPath, options);
     }
 
-    const file = fs.createReadStream(exportFilePath);
+    const file = fs.createReadStream(filePath);
     res.set({
       'Content-Type':
         format === 'json' ? 'application/json' : 'application/octet-stream',
