+++ packages/api2/src/praise/services/praise.service.ts
@@ -1,35 +1,22 @@
-import * as fs from 'fs';
-import * as csv from 'fast-csv';
 import { InjectModel } from '@nestjs/mongoose';
 import { Types } from 'mongoose';
-import {
-  PraiseModel,
-  Praise,
-  PraiseDocument,
-  PraiseExportSqlSchema,
-} from './schemas/praise.schema';
+import { PraiseModel, Praise, PraiseDocument } from '../schemas/praise.schema';
 import { ServiceException } from '@/shared/exceptions/service-exception';
 import { PeriodStatusType } from '@/periods/enums/status-type.enum';
 import { SettingsService } from '@/settings/settings.service';
 import { QuantificationsService } from '@/quantifications/services/quantifications.service';
-import { PraisePaginatedQueryDto } from './dto/praise-paginated-query.dto';
+import { PraisePaginatedQueryDto } from '../dto/praise-paginated-query.dto';
 import { Pagination } from 'mongoose-paginate-ts';
-import { EventLogService } from '../event-log/event-log.service';
+import { EventLogService } from '../../event-log/event-log.service';
 import { EventLogTypeKey } from '@/event-log/enums/event-log-type-key';
 import { QuantifyInputDto } from '@/praise/dto/quantify-input.dto';
 import { RequestContext } from 'nestjs-request-context';
 import { RequestWithAuthContext } from '@/auth/interfaces/request-with-auth-context.interface';
-import { PraisePaginatedResponseDto } from './dto/praise-paginated-response.dto';
+import { PraisePaginatedResponseDto } from '../dto/praise-paginated-response.dto';
 import { Period, PeriodModel } from '@/periods/schemas/periods.schema';
 import { Inject, Injectable, forwardRef } from '@nestjs/common';
 import { PeriodsService } from '@/periods/services/periods.service';
-import { parse } from 'json2csv';
 import { PeriodDateRangeDto } from '@/periods/dto/period-date-range.dto';
-import { ExportInputDto } from '@/shared/dto/export-input.dto';
-import { allExportsDirPath } from '@/shared/fs.shared';
-import { exec } from '@/shared/duckdb.shared';
-import duckdb from 'duckdb';
-import crypto from 'crypto';
 @Injectable()
 export class PraiseService {
   constructor(
@@ -102,154 +89,6 @@ export class PraiseService {
     return praisePagination;
   }
 
-  async generateCsvExport(options: ExportInputDto) {
-    const { periodId, startDate, endDate } = options;
-    const query = {} as any;
-
-    if (periodId) {
-      if (startDate || endDate) {
-        // If periodId is set, startDate and endDate should not be set
-        throw new ServiceException(
-          'Invalid date filtering option. When periodId is set, startDate and endDate should not be set.',
-        );
-      }
-      const period = await this.periodService.findOneById(periodId);
-      query.createdAt = await this.periodService.getPeriodDateRangeQuery(
-        period,
-      );
-    } else {
-      if (startDate && endDate) {
-        // If periodId is not set but startDate and endDate are set, use them to filter
-        query.createdAt = {
-          $gte: startDate,
-          $lte: endDate,
-        };
-      } else if (startDate || endDate) {
-        // If periodId is not set and only one of startDate and endDate is set, throw an error
-        throw new ServiceException(
-          'Invalid date filtering option. When periodId is not set, both startDate and endDate should be set.',
-        );
-      }
-    }
-
-    // Fields to include in the csv
-    const includeFields = [
-      '_id',
-      'giver',
-      'forwarder',
-      'receiver',
-      'reason',
-      'reasonRaw',
-      'score',
-      'sourceId',
-      'sourceName',
-      'createdAt',
-      'updatedAt',
-    ];
-
-    // Serialization rules
-    const transform = (doc: PraiseDocument) => ({
-      _id: doc._id,
-      giver: doc.giver._id,
-      forwarder: doc.forwarder ? doc.forwarder._id : null,
-      receiver: doc.receiver._id,
-      reason: doc.reason,
-      reasonRaw: doc.reasonRaw,
-      score: doc.score,
-      sourceId: doc.sourceId,
-      sourceName: doc.sourceName,
-      createdAt: doc.createdAt.toISOString(),
-      updatedAt: doc.updatedAt.toISOString(),
-    });
-
-    const exportDirName = await this.getExportDirName();
-    const exportId = this.getExportId(options);
-    const exportDirPath = `${allExportsDirPath}/praise/${exportDirName}/${exportId}`;
-
-    // Create the export folder if it doesn't exist
-    if (!fs.existsSync(exportDirPath)) {
-      fs.mkdirSync(exportDirPath, { recursive: true });
-    }
-
-    // Return a promise that resolves when the csv is done
-    return new Promise(async (resolve) => {
-      // Count the number of documents that match the query and write an empty csv, headers only, if there are none
-      const count = await this.praiseModel.countDocuments(query);
-      if (count === 0) {
-        fs.writeFileSync(
-          `${exportDirPath}/praise.csv`,
-          includeFields.join(','),
-        );
-        resolve(true);
-        return;
-      }
-
-      const cursor = this.praiseModel
-        .find(query)
-        .select(includeFields.join(' '))
-        .cursor();
-
-      // Create a csv writer that transforms the data using our rules
-      const csvWriter = csv.format({
-        headers: true,
-        transform,
-      });
-
-      // Pipe the csvWriter to a file
-      csvWriter.pipe(fs.createWriteStream(`${exportDirPath}/praise.csv`));
-
-      // Resolve promise when csvWriter is done
-      csvWriter.on('end', () => {
-        resolve(true);
-      });
-
-      // Pipe the cursor to the csvWriter
-      cursor.pipe(csvWriter);
-    });
-  }
-
-  /**
-   * Generates all export files - csv and parquet
-   */
-  async generateAllExports(options: ExportInputDto) {
-    const exportDirName = await this.getExportDirName();
-    const exportId = this.getExportId(options);
-    const exportDirPath = `${allExportsDirPath}/praise/${exportDirName}/${exportId}`;
-
-    await this.generateCsvExport(options);
-
-    // Create a duckdb database, import the csv file, and export it to parquet
-    const db = new duckdb.Database(':memory:');
-    await exec(db, `CREATE TABLE praise (${PraiseExportSqlSchema})`);
-    await exec(
-      db,
-      `COPY praise FROM '${exportDirPath}/praise.csv' (AUTO_DETECT TRUE, HEADER TRUE);`,
-    );
-    await exec(
-      db,
-      `COPY praise TO '${exportDirPath}/praise.parquet' (FORMAT PARQUET, COMPRESSION ZSTD);`,
-    );
-  }
-
-  /**
-   * The export directory name is the _id of the last inserted document
-   */
-  async getExportDirName(): Promise<string> {
-    const latestAdded = await this.findLatestAdded();
-    return latestAdded._id.toString();
-  }
-
-  /**
-   *  Create a hashed id based on the export options excluding export format
-   */
-  getExportId(options: ExportInputDto): string {
-    const { periodId, startDate, endDate } = options;
-    return crypto
-      .createHash('sha256')
-      .update(JSON.stringify({ periodId, startDate, endDate }))
-      .digest('hex');
-  }
-
   /**
    * Find one praise by id
    * @param _id
@@ -285,7 +124,7 @@ export class PraiseService {
   /**
    * Find the lastest added praise
    */
-  async findLatestAdded(): Promise<Praise> {
+  async findLatest(): Promise<Praise> {
     const praise = await this.praiseModel
       .find()
       .limit(1)
